+++ apix-interface/src/main/java/eu/xenit/apix/data/NodeRef.java
@@ -2,6 +2,7 @@ package eu.xenit.apix.data;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
+
 import java.io.Serializable;
 
 /**
@@ -14,8 +15,10 @@ public class NodeRef implements Serializable {
 
     private static final String FORWARD_SLASH = "/";
     private static final String SPACESTORE_DIVIDER = "://";
-
     private String value;
+
+    public NodeRef() {
+    }
 
     @JsonCreator
     public NodeRef(String s) {