+++ apix-rest-v1/src/main/java/eu/xenit/apix/rest/v1/dictionary/DictionaryWebScript1.java
@@ -1,13 +1,8 @@
 package eu.xenit.apix.rest.v1.dictionary;
 
-import com.github.dynamicextensionsalfresco.webscripts.annotations.Authentication;
-import com.github.dynamicextensionsalfresco.webscripts.annotations.AuthenticationType;
-import com.github.dynamicextensionsalfresco.webscripts.annotations.FormatStyle;
-import com.github.dynamicextensionsalfresco.webscripts.annotations.HttpMethod;
-import com.github.dynamicextensionsalfresco.webscripts.annotations.RequestParam;
-import com.github.dynamicextensionsalfresco.webscripts.annotations.Uri;
-import com.github.dynamicextensionsalfresco.webscripts.annotations.UriVariable;
-import com.github.dynamicextensionsalfresco.webscripts.annotations.WebScript;
+import com.gradecak.alfresco.mvc.annotation.AlfrescoAuthentication;
+import com.gradecak.alfresco.mvc.annotation.AlfrescoTransaction;
+import com.gradecak.alfresco.mvc.annotation.AuthenticationType;
 import eu.xenit.apix.data.QName;
 import eu.xenit.apix.dictionary.IDictionaryService;
 import eu.xenit.apix.dictionary.aspects.AspectDefinition;
@@ -18,109 +13,93 @@ import eu.xenit.apix.dictionary.types.Types;
 import eu.xenit.apix.properties.Properties;
 import eu.xenit.apix.properties.PropertyDefinition;
 import eu.xenit.apix.rest.v1.ApixV1Webscript;
-import eu.xenit.apix.rest.v1.RestV1Config;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
-import java.io.IOException;
-import org.apache.http.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.extensions.webscripts.WebScriptResponse;
-import org.springframework.stereotype.Component;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-@WebScript(baseUri = RestV1Config.BaseUrl + "/dictionary", families = RestV1Config.Family, defaultFormat = "json",
-        description = "Retrieves Dictionary information", value = "Dictionary")
-@Authentication(AuthenticationType.USER)
-@Component("eu.xenit.apix.rest.v1.property.DictionaryWebScript1")
+import javax.servlet.http.HttpServletRequest;
+
+@AlfrescoAuthentication(AuthenticationType.USER)
+@RestController
 public class DictionaryWebScript1 extends ApixV1Webscript {
 
-    Logger logger = LoggerFactory.getLogger(DictionaryWebScript1.class);
-    @Autowired
-    IDictionaryService dictionaryService;
+    private static final Logger logger = LoggerFactory.getLogger(DictionaryWebScript1.class);
+    private final IDictionaryService dictionaryService;
 
+    public DictionaryWebScript1(
+            @Qualifier(("eu.xenit.apix.dictionary.IDictionaryService")) IDictionaryService dictionaryService) {
+        this.dictionaryService = dictionaryService;
+    }
 
-    @Uri(value = "/properties/{qname}", method = HttpMethod.GET, formatStyle = FormatStyle.ARGUMENT)
-    @ApiOperation(value = "Return the definition of a property", notes = "")
-    @ApiResponses(@ApiResponse(code = 200, message = "Success", response = PropertyDefinition.class))
-    public void getPropertyDefinition(@UriVariable final String qname, WebScriptResponse webScriptResponse)
-            throws IOException {
-        eu.xenit.apix.data.QName apixQName = new eu.xenit.apix.data.QName(qname);
-        PropertyDefinition propDef = dictionaryService.GetPropertyDefinition(apixQName);
+    @AlfrescoTransaction(readOnly = true)
+    @GetMapping(value = "/v1/dictionary/properties/**")
+    public ResponseEntity<?> getPropertyDefinition(HttpServletRequest request) {
+        QName qname = extractQNameFromUrlPath(request, "/v1/dictionary/properties/");
+        PropertyDefinition propDef = dictionaryService.GetPropertyDefinition(qname);
         if (propDef == null) {
-            webScriptResponse.setStatus(HttpStatus.SC_NOT_FOUND);
+            return ResponseEntity.notFound().build();
         }
-        writeJsonResponse(webScriptResponse, propDef);
+        return writeJsonResponse(propDef);
     }
 
-
-    @Uri(value = "/properties", method = HttpMethod.GET, formatStyle = FormatStyle.ARGUMENT)
-    @ApiOperation(value = "Return properties", notes = "")
-    @ApiResponses(@ApiResponse(code = 200, message = "Success", response = Properties.class))
-    public void getProperties(
-            WebScriptResponse webScriptResponse) throws IOException {
-        Properties properties = dictionaryService.getProperties();
-        writeJsonResponse(webScriptResponse, properties);
+    @AlfrescoTransaction(readOnly = true)
+    @GetMapping(value = "/v1/dictionary/properties")
+    public ResponseEntity<Properties> getProperties() {
+        return writeJsonResponse(dictionaryService.getProperties());
     }
 
-
-    @Uri(value = "/types", method = HttpMethod.GET, formatStyle = FormatStyle.ARGUMENT)
-    @ApiOperation(value = "Return the definitions of types", notes = "")
-    @ApiResponses(@ApiResponse(code = 200, message = "Success", response = Types.class))
-    public void getSubTypeDefinitions(@RequestParam(defaultValue = "sys:base", required = false) final String parent,
-            WebScriptResponse webScriptResponse) throws IOException {
-        QName apixQName = new QName(parent);
-        Types types = dictionaryService.GetSubTypeDefinitions(apixQName, true);
-        writeJsonResponse(webScriptResponse, types);
+    @AlfrescoTransaction(readOnly = true)
+    @GetMapping(value = "/v1/dictionary/types")
+    public ResponseEntity<Types> getSubTypeDefinitions(@RequestParam(defaultValue = "sys:base", required = false) final String parent) {
+        return writeJsonResponse(
+                dictionaryService.GetSubTypeDefinitions(
+                        new QName(parent), true
+                )
+        );
     }
 
-    @Uri(value = "/types/{qname}", method = HttpMethod.GET, formatStyle = FormatStyle.ARGUMENT)
-    @ApiOperation(value = "Return the definition of a type", notes = "")
-    @ApiResponses({
-            @ApiResponse(code = 200, message = "Success", response = TypeDefinition.class),
-            @ApiResponse(code = 404, message = "Not Found")})
-    public void getTypeDefinition(@UriVariable final String qname, WebScriptResponse webScriptResponse)
-            throws IOException {
-        logger.debug("Received type qname %s", qname);
-        eu.xenit.apix.data.QName apixQName = new eu.xenit.apix.data.QName(qname);
-        TypeDefinition classDef = dictionaryService.GetTypeDefinition(apixQName);
+    @AlfrescoTransaction(readOnly = true)
+    @GetMapping(value = "/v1/dictionary/types/**")
+    public ResponseEntity<?> getTypeDefinition(HttpServletRequest request) {
+        QName qname = extractQNameFromUrlPath(request, "/v1/dictionary/types/");
+        logger.debug("Received type qname {}", qname);
+        TypeDefinition classDef = dictionaryService.GetTypeDefinition(qname);
         if (classDef == null) {
-            webScriptResponse.setStatus(HttpStatus.SC_NOT_FOUND);
+            return ResponseEntity.notFound().build();
         }
-        writeJsonResponse(webScriptResponse, classDef);
+        return writeJsonResponse(classDef);
     }
 
-    @Uri(value = "/aspects/{qname}", method = HttpMethod.GET, formatStyle = FormatStyle.ARGUMENT)
-    @ApiOperation(value = "Return the definition of a aspect", notes = "")
-    @ApiResponses(@ApiResponse(code = 200, message = "Success", response = AspectDefinition.class))
-    public void getAspectDefinition(@UriVariable final String qname, WebScriptResponse webScriptResponse)
-            throws IOException {
-        logger.debug("Received aspect qname %s", qname);
-        eu.xenit.apix.data.QName apixQName = new eu.xenit.apix.data.QName(qname);
-        AspectDefinition classDef = dictionaryService.GetAspectDefinition(apixQName);
+    @AlfrescoTransaction(readOnly = true)
+    @GetMapping(value = "/v1/dictionary/aspects/**")
+    public ResponseEntity<?> getAspectDefinition(HttpServletRequest request) {
+        QName qname = extractQNameFromUrlPath(request, "/v1/dictionary/aspects/");
+        logger.debug("Received aspect qname {}", qname);
+        AspectDefinition classDef = dictionaryService.GetAspectDefinition(qname);
         if (classDef == null) {
-            webScriptResponse.setStatus(HttpStatus.SC_NOT_FOUND);
+            return ResponseEntity.notFound().build();
         }
-        writeJsonResponse(webScriptResponse, classDef);
+        return writeJsonResponse(classDef);
     }
 
-
-    @Uri(value = "/aspects", method = HttpMethod.GET, formatStyle = FormatStyle.ARGUMENT)
-    @ApiOperation(value = "Return apects", notes = "")
-    @ApiResponses(@ApiResponse(code = 200, message = "Success", response = Aspects.class))
-    public void getAspects(WebScriptResponse webScriptResponse) throws IOException {
-        Aspects aspects = dictionaryService.getAspects();
-        writeJsonResponse(webScriptResponse, aspects);
+    @AlfrescoTransaction(readOnly = true)
+    @GetMapping(value = "/v1/dictionary/aspects")
+    public ResponseEntity<Aspects> getAspects() {
+        return writeJsonResponse(dictionaryService.getAspects());
     }
 
-    @Uri(value = "/namespaces", method = HttpMethod.GET, formatStyle = FormatStyle.ARGUMENT)
-    @ApiOperation(value = "Returns the namespaces", notes = "")
-    @ApiResponses(@ApiResponse(code = 200, message = "Success", response = Namespaces.class))
-    public void getNamespaces(WebScriptResponse webScriptResponse) throws IOException {
-        Namespaces namespaces = dictionaryService.getNamespaces();
-        writeJsonResponse(webScriptResponse, namespaces);
+    @AlfrescoTransaction(readOnly = true)
+    @GetMapping(value = "/v1/dictionary/namespaces")
+    public ResponseEntity<Namespaces> getNamespaces() {
+        return writeJsonResponse(dictionaryService.getNamespaces());
     }
 
-
+    private QName extractQNameFromUrlPath(HttpServletRequest request, String path) {
+        String qnameValue = request.getRequestURI().split(request.getContextPath() + path)[1];
+        return new QName(qnameValue);
+    }
 }