+++ apix-rest-v1/src/main/java/eu/xenit/apix/rest/v1/nodes/InheritFromParent.java
@@ -2,12 +2,14 @@ package eu.xenit.apix.rest.v1.nodes;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.annotations.ApiModelProperty;
 
 class InheritFromParent {
 
-    @ApiModelProperty(required = true)
-    public boolean inheritFromParent;
+    private boolean inheritFromParent;
+
+    public InheritFromParent() {
+
+    }
 
     @JsonCreator
     public InheritFromParent(@JsonProperty("inheritFromParent") boolean inheritFromParent) {
@@ -15,11 +17,10 @@ class InheritFromParent {
     }
 
     public boolean isInheritFromParent() {
-        return this.inheritFromParent;
+        return inheritFromParent;
     }
 
     public void setInheritFromParent(boolean inheritFromParent) {
         this.inheritFromParent = inheritFromParent;
     }
-
 }