+++ apix-rest-v1/src/main/java/eu/xenit/apix/rest/v1/temp/LogsWebscript.java
@@ -1,48 +1,39 @@
 package eu.xenit.apix.rest.v1.temp;
 
-import com.github.dynamicextensionsalfresco.webscripts.annotations.Authentication;
-import com.github.dynamicextensionsalfresco.webscripts.annotations.AuthenticationType;
-import com.github.dynamicextensionsalfresco.webscripts.annotations.RequestParam;
-import com.github.dynamicextensionsalfresco.webscripts.annotations.Uri;
-import com.github.dynamicextensionsalfresco.webscripts.annotations.WebScript;
+import com.gradecak.alfresco.mvc.annotation.AlfrescoAuthentication;
+import com.gradecak.alfresco.mvc.annotation.AlfrescoTransaction;
+import com.gradecak.alfresco.mvc.annotation.AuthenticationType;
 import eu.xenit.apix.rest.v1.ApixV1Webscript;
-import eu.xenit.apix.rest.v1.RestV1Config;
-import io.swagger.annotations.Api;
 import java.io.File;
 import java.io.IOException;
-import java.io.Writer;
-import java.util.ArrayList;
-import org.springframework.extensions.webscripts.WebScriptResponse;
-import org.springframework.stereotype.Component;
+import org.springframework.core.env.Environment;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 
-/**
- * Created by Michiel Huygen on 03/05/2016.
- */
-@WebScript(baseUri = RestV1Config.BaseUrl, families = RestV1Config.Family, defaultFormat = "json",
-        description = "Display logs")
-@Api(hidden = true)
-@Component("eu.xenit.apix.rest.v1.temp.LogsWebscript")
+@RestController
 public class LogsWebscript extends ApixV1Webscript {
 
-    @Uri(value = "/tmp/log", defaultFormat = "text")
-    @Authentication(AuthenticationType.ADMIN)
-    public void showLog(@RequestParam(defaultValue = "200") int lines, WebScriptResponse resp) throws IOException {
-        ArrayList<String> output = new ArrayList<String>();
+    private final String logPath;
 
-        try (ReversedLinesFileReader reader = new ReversedLinesFileReader(
-                new File("/opt/alfresco/tomcat/logs/catalina.out"))) {
+    public LogsWebscript(Environment env) {
+        logPath = env.resolvePlaceholders("$CATALINA_HOME/logs/catalina.out");
+    }
+    @AlfrescoTransaction(readOnly = true)
+    @GetMapping(value = "/v1/tmp/log", produces = { MediaType.TEXT_PLAIN_VALUE })
+    @AlfrescoAuthentication(AuthenticationType.ADMIN)
+    public ResponseEntity<String> showLog(@RequestParam(defaultValue = "200") int lines) throws IOException {
+        StringBuilder log = new StringBuilder();
+        File logFile = new File(logPath);
+        try (ReversedLinesFileReader reader = new ReversedLinesFileReader(logFile)) {
             for (int i = 0; i < lines; i++) {
-                output.add(reader.readLine() + "\n");
+                log.append(reader.readLine())
+                    .append("\n");
             }
         }
-
-        Writer writer = resp.getWriter();
-
-        for (int i = lines - 1; i >= 0; i--) {
-            writer.append(output.get(i));
-        }
-
-
+        return ResponseEntity.ok(log.toString());
     }
 }