+++ apix-rest-v1/src/main/java/eu/xenit/apix/rest/v2/groups/SetUsersInGroupOptions.java
@@ -2,12 +2,13 @@ package eu.xenit.apix.rest.v2.groups;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.annotations.ApiModelProperty;
 
 class SetUsersInGroupOptions {
 
-    @ApiModelProperty(required = true)
     private String[] users;
+
+    public SetUsersInGroupOptions() {
+    }
 
     @JsonCreator
     public SetUsersInGroupOptions(@JsonProperty("users") String[] users) {